requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.31.0
python-multipart==0.0.9
fastembed==0.6.0
qdrant-client==1.13.3
python-dotenv==1.0.1
torch
transformers
Pillow
ftfy
regex
tqdm
numpy